aibotsallinone/requirements.txt
fastapi
uvicorn
anthropic
openai
google-generativeai
python-dotenv
pydantic